Fix conflicting types for exec in command.c

+++ src/command.c
@@ -1,7 +1,7 @@
 #include "command.h"
 
 /* Executes a command, and prints the output of the command */
-char* exec(const char* command, size_t size) {
+char* exec(const char* command, off_t size) {
   FILE *fp;
   char buffer[size];
